requirements.txt
python-telegram-bot[job-queue]==20.7
Pillow==10.2.0
aiohttp==3.9.3
python-dotenv==1.0.1
py-clob-client==0.28.0
web3==7.6.1
pytest==8.4.2
pytest-asyncio==0.24.0
psycopg2-binary==2.9.9
google-cloud-kms==2.20.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
requests==2.31.0
loguru==0.7.2
